feat: better modularity in agent conversation history context
