+++ frontend/lib/chat/agents/pipe.ts
@@ -39,13 +39,8 @@ export class PipeAgent extends BaseAgent<PipeAgentInput, PipeOutput> {
   protected getTools(input: PipeAgentInput): Record<string, any> {
     return input.tools
   }
-  
-  protected override getConversationHistory(input: PipeAgentInput) {
-    const userMessages = input.messages.filter((m) => m.role === 'user')
-    if (userMessages.length > 1) {
-      return JSON.stringify(userMessages.slice(0, -1), null, 2)
-    }
 
+  protected generateConversationHistoryReceipt(_input: PipeAgentInput): string {
     return ''
   }
